scripts/requirements.txt
pypdf>=5.0.0
python-dotenv>=1.0.0
